feat(issues): add contact information redirects
Sometimes, we receive feedback on Codeberg and on its policies, as
well as even spam reports. The repository is primarily intended
for legal documents (as well as potential inconsistencies/problems
that they, or their translations, may have). Feedback on the policies
and their respective implementations is off-topic.

It makes sense to redirect people to Codeberg/Community and, should
a matter of discussion need discretion, our Contact page.